requirements.txt
absl-py @ file:///work/perseverance-python-buildout/croot/absl-py_1728398107234/work
actionlib==1.14.0
angles==1.9.13
ansitable==0.11.4
base_local_planner==1.17.3
Brotli @ file:///croot/brotli-split_1736182456865/work
catkin==0.8.10
-e git+ssh://git@github.com/Shono1/rbe575project.git@b05ea3039292d22d427ed18c6bae722ccccd23e4#egg=cbf_toolbox
cbfpy==0.0.1
cfgv==3.4.0
colored==2.3.0
contourpy @ file:///croot/contourpy_1732540045555/work
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
diagnostic-updater==1.11.0
dill==0.4.0
distlib==0.3.9
dynamic-reconfigure==1.7.3
dynamixel_sdk==3.7.51
filelock==3.18.0
fonttools @ file:///croot/fonttools_1737039080035/work
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gurobipy==12.0.1
identify==2.6.9
interactive-markers==1.12.0
jax @ file:///io/buildout/croot/jax_1735870278957/work
jaxlib @ file:///io/buildout/croot/jaxlib_1735845688747/work/dist/jaxlib-0.4.35-cp313-cp313-manylinux2014_x86_64.whl#sha256=cb18e9b48c48cf01e6e3bc433bd44f738860acb0797b5b52bb05006199bbdb11
jaxopt @ file:///home/conda/feedstock_root/build_artifacts/jaxopt_1744618047935/work
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
kiwisolver @ file:///croot/kiwisolver_1737039087198/work
laser_geometry==1.6.7
matplotlib==3.10.0
mavros==1.20.0
message-filters==1.16.0
ml-dtypes==0.4.1
moveit-commander==1.1.16
moveit-core==1.1.16
moveit-python==0.4.5
moveit-ros-planning-interface==1.1.16
moveit-ros-visualization==1.1.16
moveit-task-constructor-core==0.1.3
mpmath==1.3.0
nodeenv==1.9.1
numpy==1.26.4
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging @ file:///croot/packaging_1734472117206/work
pgraph-python==0.6.3
pillow @ file:///croot/pillow_1738010226202/work
platformdirs==4.3.7
pr2_power_board==1.1.10
pre_commit==4.2.0
progress==1.6
py-trees==0.7.6
pyparsing @ file:///io/buildout/croot/pyparsing_1735862672222/work
PyQt6==6.7.1
PyQt6_sip @ file:///croot/pyqt-split_1740498191142/work/pyqt_sip
python-dateutil @ file:///work/perseverance-python-buildout/croot/python-dateutil_1728385355105/work
python-qt-binding==0.4.4
PyYAML==6.0.2
qpax==0.0.9
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
resource_retriever==1.12.7
roboticstoolbox-python==1.1.1
rosbag==1.16.0
rosclean==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosserial_arduino==0.9.2
rosserial_client==0.9.2
rosserial_python==0.9.2
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-console==0.4.12
rqt-controller-manager==0.20.0
rqt-ez-publisher==0.6.1
rqt-ground-robot-teleop==1.4.2
rqt-joint-trajectory-controller==0.21.2
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-multiplot==0.0.12
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_pr2_dashboard==0.4.0
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_py_trees==0.4.1
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rtb-data==1.0.1
rviz==1.14.25
scipy @ file:///croot/scipy_1743153247252/work/dist/scipy-1.15.2-cp313-cp313-linux_x86_64.whl#sha256=bb1d0c888ef39ba5effc7c1fe69c5d342fc6b913441a954dc2856f07ccb9b768
sensor-msgs==1.13.1
setuptools==72.1.0
sip @ file:///croot/sip_1738856193618/work
six @ file:///croot/six_1744271502820/work
smach==2.5.2
smach-ros==2.5.2
smach-viewer==4.1.0
smclib==1.8.6
spatialgeometry==1.1.0
spatialmath-python==1.1.14
srdfdom==0.6.4
swift-sim==1.1.0
sympy==1.13.3
tf==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
topic-tools==1.16.0
tornado @ file:///io/buildout/croot/tornado_1735842129908/work
typing_extensions==4.13.2
unique_id==1.0.6
urdfdom-py==0.4.6
virtualenv==20.30.0
websockets==15.0.1
wheel==0.45.1
xacro==1.14.17
